wi https://github.com/metasfresh/metasfresh/issues/11525
