+++ backend/de.metas.payment.esr/src/main/java/de/metas/payment/esr/api/impl/ESRImportDAO.java
@@ -99,7 +99,7 @@ public class ESRImportDAO implements IESRImportDAO
 	{
 		saveRecord(esrImportLine);
 	}
-	
+
 	@Override
 	public List<I_ESR_Import> getByIds(@NonNull final Set<ESRImportId> esrImportIds)
 	{
